Reduce bundle size
