+++ packages/rehype-mathjax/renderer/svg.js
@@ -1,0 +1,19 @@
+const {mathjax} = require('mathjax-full/js/mathjax')
+const {TeX} = require('mathjax-full/js/input/tex')
+const {SVG} = require('mathjax-full/js/output/svg')
+const {AllPackages} = require('mathjax-full/js/input/tex/AllPackages')
+const AbstractRenderer = require('./node')
+
+class SVGRenderer extends AbstractRenderer {
+  constructor(options) {
+    super()
+    this.InputJax = new TeX({packages: AllPackages})
+    this.OutputJax = new SVG(options)
+    this.mathDocument = mathjax.document('', {
+      InputJax: this.InputJax,
+      OutputJax: this.OutputJax
+    })
+  }
+}
+
+module.exports = SVGRenderer
